backend/requirements.txt
fastapi>=0.105.0
starlette==0.31.0
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
onnxruntime==1.16.3
scipy==1.11.4
tqdm==4.66.1
aiohttp==3.9.1
filetype==1.2.0
pooch==1.8.0